+++ src/components/categories-filter/CategoriesFilter.component.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import appStyles from '@/App.module.css';
 import { CategoriesDataProviderComponent } from '@/data-providers/CategoriesDataProvider.component.tsx';
-import { transformCheckedCategories, transformCheckedCategoryIds } from '@/helpers/CategoriesFilterHelper.ts';
+import { transformCheckedCategories, transformCheckedCategoryIds } from '@/helpers/CategoriesFilter.helper.ts';
 import type { CheckedCategories } from '@/types/CheckedCategories.type.ts';
 
 import styles from './categories-filter.module.css';
